generator ranges commit

+++ SatteliteSimulator/Controller/Controller.h
@@ -19,6 +19,6 @@ namespace controller {
 		Sample GetSample();
 		void ControlRotation(Vector3d rotSpeed);
 		void ControlInstanceSpeed(Vector3d instSpeed);
-		void Generate(int rangeStart, int rangeEnd, int sampleSize);
+		void Generate(double rangeStart, double rangeEnd, int sampleSize);
 	};
 }
